Code Cleaning and more
-Cleaned code
-Made code more readable
-Filled in info on the instruction screen

-Jordi Prummel

+++ I.D.K/Assets/Scripts/Character/PlayerLives.cs
@@ -2,8 +2,8 @@
 using System.Collections;
 
 public class PlayerLives : MonoBehaviour {
-	public float lives;
-	private float maxLives = 3;
+	public int lives;
+	private int maxLives = 3;
     public ParticleSystemRenderer explosion;
     public Transform firePoint;
